Fixed move class and chesspiece class as noted
